feat: create dashboard page
